moving shapes to a more logical place
